Эксперимент с разнесением классов. Неудачный.
