--- .idea/compiler.xml
@@ -1,8 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project version="4">
-  <component name="CompilerConfiguration">
-    <excludeFromCompile>
-      <file url="file://$PROJECT_DIR$/src/ru/vachok/Main.java" />
-    </excludeFromCompile>
-  </component>
-</project>
